Add `BlogList` component
